requirements.txt
litellm
duckduckgo-search
pydantic
python-dotenv
google-generativeai
